factor out common code into third feature
