+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jdk/JmxServerFeature.java
@@ -29,7 +29,6 @@ package com.oracle.svm.hosted.jdk;
 import com.oracle.svm.core.feature.InternalFeature;
 
 import java.lang.management.PlatformManagedObject;
-import java.util.Arrays;
 
 import java.util.Set;
 
@@ -40,83 +39,17 @@ import org.graalvm.nativeimage.impl.ConfigurationCondition;
 import org.graalvm.nativeimage.ImageSingletons;
 import com.oracle.svm.core.configure.ResourcesRegistry;
 import com.oracle.svm.core.jdk.proxy.DynamicProxyRegistry;
-import com.oracle.svm.core.jni.JNIRuntimeAccess;
 import org.graalvm.nativeimage.hosted.RuntimeReflection;
-import org.graalvm.nativeimage.hosted.RuntimeSerialization;
 import com.oracle.svm.core.jdk.RuntimeSupport;
 import com.oracle.svm.core.jdk.management.ManagementAgentStartupHook;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.VMInspectionOptions;
-import com.oracle.svm.util.ReflectionUtil;
 
 @AutomaticallyRegisteredFeature
 public class JmxServerFeature implements InternalFeature {
     @Override
     public boolean isInConfiguration(IsInConfigurationAccess access) {
         return VMInspectionOptions.hasJmxServerSupport();
-    }
-
-    @Override
-    public void afterRegistration(AfterRegistrationAccess access) {
-        org.graalvm.nativeimage.impl.RuntimeClassInitializationSupport rci = ImageSingletons.lookup(org.graalvm.nativeimage.impl.RuntimeClassInitializationSupport.class);
-        if (org.graalvm.compiler.serviceprovider.JavaVersionUtil.JAVA_SPEC <= 11) {
-            // Activation has been removed in JDK 17.
-            rci.initializeAtBuildTime("sun.rmi.server.Activation$ActivationSystemImpl_Stub", "JDK11 JMX support");
-        }
-        rci.initializeAtBuildTime("jdk.management.jfr.SettingDescriptorInfo", "JMX support");
-        rci.initializeAtBuildTime("sun.rmi.runtime.Log$LogStreamLog", "JMX support");
-        rci.initializeAtBuildTime("sun.rmi.runtime.Log", "JMX support");
-        rci.initializeAtBuildTime("java.rmi.server.LogStream", "JMX support");
-        rci.initializeAtBuildTime("sun.rmi.runtime.Log$LoggerLog", "JMX support");
-        rci.initializeAtBuildTime("sun.rmi.server.Util", "JMX support");
-        rci.initializeAtBuildTime("jdk.management.jfr.MBeanUtils", "JMX support");
-
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.StandardMBeanIntrospector", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.MXBeanIntrospector", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.MXBeanLookup", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.util.EnvHelp", "JMX support");
-        rci.initializeAtBuildTime("java.rmi.server.RemoteObjectInvocationHandler", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.Introspector", "JMX support");
-        rci.initializeAtBuildTime("java.beans.Introspector", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.JavaBeansAccessor", "JMX support");
-
-        rci.initializeAtBuildTime("com.sun.jmx.remote.security.JMXPluggableAuthenticator", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.mbeanserver.MBeanInstantiator", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.internal.ArrayNotificationBuffer", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.security.HashedPasswordManager", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.security.JMXSubjectDomainCombiner", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.internal.ServerCommunicatorAdmin", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.defaults.JmxProperties", "JMX support");
-        rci.initializeAtBuildTime("com.sun.jmx.remote.internal.ServerNotifForwarder", "JMX support");
-
-        rci.initializeAtRunTime("sun.rmi.transport.ConnectionInputStream", "JMX support");
-        rci.initializeAtRunTime("java.rmi.MarshalledObject$MarshalledObjectInputStream", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.UnicastRef2", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.UnicastRef", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.MarshalInputStream", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.runtime.NewThreadAction", "JMX support");
-        rci.initializeAtRunTime("com.sun.jmx.remote.security.FileLoginModule", "JMX support");
-        rci.initializeAtRunTime("com.sun.jmx.remote.security.JMXPluggableAuthenticator$FileLoginConfig", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.DGCImpl", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.DGCAckHandler", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.GC", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.DGCClient", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.ObjectTable", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.tcp.TCPEndpoint", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.tcp.TCPChannel", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.tcp.TCPTransport", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.transport.Transport", "JMX support");
-        rci.initializeAtRunTime("java.rmi.server.ObjID", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.UnicastServerRef", "JMX support");
-        rci.initializeAtRunTime("java.rmi.server.UID", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.runtime.RuntimeUtil", "JMX support");
-        rci.initializeAtRunTime("java.rmi.dgc.VMID", "JMX support");
-        rci.initializeAtRunTime("java.rmi.server.RMIClassLoader", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.LoaderHandler", "JMX support");
-        rci.initializeAtRunTime("java.rmi.server.RemoteServer", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.registry.RegistryImpl", "JMX support");
-        rci.initializeAtRunTime("java.rmi.server.UnicastRemoteObject", "JMX support");
-        rci.initializeAtRunTime("sun.rmi.server.UnicastServerRef2", "JMX support");
     }
 
     private static void handleNativeLibraries(BeforeAnalysisAccess access) {
@@ -135,8 +68,6 @@ public class JmxServerFeature implements InternalFeature {
     public void beforeAnalysis(BeforeAnalysisAccess access) {
         handleNativeLibraries(access);
         registerJMXAgentResources();
-        configureJNI();
-        configureSerialization(access);
         configureReflection(access);
         configureProxy(access);
         RuntimeSupport.getRuntimeSupport().addStartupHook(new ManagementAgentStartupHook());
@@ -159,84 +90,6 @@ public class JmxServerFeature implements InternalFeature {
                         access.findClassByName("java.rmi.registry.Registry"));
 
         dynamicProxySupport.addProxyClass(access.findClassByName("javax.management.remote.rmi.RMIServer"));
-
-        dynamicProxySupport.addProxyClass(access.findClassByName("jdk.management.jfr.FlightRecorderMXBean"),
-                        access.findClassByName("javax.management.NotificationEmitter"));
-
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.RuntimeMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.ClassLoadingMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("com.sun.management.ThreadMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.ThreadMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.GarbageCollectorMXBean"), access.findClassByName("javax.management.NotificationEmitter"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("com.sun.management.GarbageCollectorMXBean"), access.findClassByName("javax.management.NotificationEmitter"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("com.sun.management.OperatingSystemMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.OperatingSystemMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.MemoryManagerMXBean"), access.findClassByName("javax.management.NotificationEmitter"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.BufferPoolMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.MemoryPoolMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("java.lang.management.MemoryMXBean"), access.findClassByName("javax.management.NotificationEmitter"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("com.sun.management.UnixOperatingSystemMXBean"));
-        dynamicProxySupport.addProxyClass(access.findClassByName("com.sun.management.java.lang.management.CompilationMXBean"));
-    }
-
-    private static void configureJNI() {
-        JNIRuntimeAccess.register(Arrays.class);
-        JNIRuntimeAccess.register(ReflectionUtil.lookupMethod(Arrays.class, "asList", Object[].class));
-    }
-
-    private static void configureSerialization(BeforeAnalysisAccess access) {
-        String[] classes = new String[]{
-                        "[B", "com.oracle.svm.core.jdk.UnsupportedFeatureError",
-                        "java.io.IOException", "java.lang.Boolean", "java.lang.ClassCastException", "java.lang.Error",
-                        "java.lang.Exception", "java.lang.IllegalArgumentException", "java.lang.IllegalStateException",
-                        "java.lang.Integer", "[Ljava.lang.Integer;", "java.lang.Long", "java.lang.NoSuchMethodException",
-                        "java.lang.Number", "[Ljava.lang.Object;", "java.lang.ReflectiveOperationException",
-                        "java.lang.RuntimeException", "java.lang.SecurityException", "java.lang.StackTraceElement",
-                        "[Ljava.lang.StackTraceElement;", "java.lang.String", "java.lang.Throwable",
-                        "java.lang.UnsupportedOperationException",
-                        "java.rmi.MarshalledObject", "[Ljava.rmi.MarshalledObject;", "java.rmi.RemoteException",
-                        "java.rmi.ServerError", "java.rmi.dgc.Lease", "java.rmi.dgc.VMID", "java.rmi.server.ObjID",
-                        "[Ljava.rmi.server.ObjID;", "java.rmi.server.RemoteObject", "java.rmi.server.RemoteStub",
-                        "java.rmi.server.UID", "java.security.GeneralSecurityException", "java.util.ArrayList",
-                        "java.util.Arrays$ArrayList", "java.util.Collections$EmptyList",
-                        "java.util.Collections$UnmodifiableCollection", "java.util.Collections$UnmodifiableList",
-                        "java.util.Collections$UnmodifiableRandomAccessList",
-                        "java.util.EventObject", // Required for notifications
-                        "java.util.HashMap", "java.util.HashSet",
-                        "java.util.LinkedHashMap", "java.util.MissingResourceException", "java.util.TreeMap",
-                        "java.util.Vector", "javax.management.Attribute", "javax.management.AttributeList",
-                        "javax.management.AttributeNotFoundException",
-                        "javax.management.AttributeChangeNotification", // Required for
-                                                                        // notifications
-                        "javax.management.InstanceNotFoundException",
-                        "javax.management.JMException", "javax.management.MBeanAttributeInfo",
-                        "[Ljavax.management.MBeanAttributeInfo;", "[Ljavax.management.MBeanConstructorInfo;",
-                        "javax.management.MBeanFeatureInfo", "javax.management.MBeanInfo",
-                        "javax.management.MBeanNotificationInfo", "[Ljavax.management.MBeanNotificationInfo;",
-                        "javax.management.MBeanOperationInfo", "[Ljavax.management.MBeanOperationInfo;",
-                        "javax.management.MBeanParameterInfo", "[Ljavax.management.MBeanParameterInfo;",
-                        "javax.management.Notification", // Required for notifications
-                        "javax.management.NotificationFilterSupport", "javax.management.ObjectName",
-                        "[Ljavax.management.ObjectName;", "javax.management.OperationsException", "javax.management.ReflectionException",
-                        "javax.management.openmbean.ArrayType", "javax.management.openmbean.CompositeDataSupport",
-                        "[Ljavax.management.openmbean.CompositeData;", "javax.management.openmbean.CompositeType",
-                        "javax.management.openmbean.OpenMBeanAttributeInfoSupport",
-                        "javax.management.openmbean.OpenMBeanParameterInfoSupport", "javax.management.openmbean.OpenType",
-                        "javax.management.openmbean.SimpleType", "javax.management.openmbean.TabularDataSupport",
-                        "javax.management.openmbean.TabularType", "javax.management.remote.NotificationResult",
-                        "javax.management.remote.TargetedNotification", // Required for
-                                                                        // notifications
-                        "[Ljavax.management.remote.TargetedNotification;", "javax.management.remote.rmi.RMIConnectionImpl_Stub",
-                        "javax.management.remote.rmi.RMIServerImpl_Stub", "javax.management.RuntimeMBeanException",
-                        "javax.rmi.ssl.SslRMIClientSocketFactory", // Required for SSL
-                        "[Ljavax.security.auth.Subject;", "javax.security.auth.login.FailedLoginException",
-                        "javax.security.auth.login.LoginException", "[J", "java.rmi.NoSuchObjectException",
-                        "javax.management.JMRuntimeException", "javax.management.RuntimeErrorException"
-        };
-
-        for (String clazz : classes) {
-            RuntimeSerialization.register(access.findClassByName(clazz));
-        }
     }
 
     private static void configureReflection(BeforeAnalysisAccess access) {
@@ -260,64 +113,25 @@ public class JmxServerFeature implements InternalFeature {
 
         // Only JmxServerFeature, not JmxClientFeature, has registrations for platform MBeans
         String[] classes = new String[]{
-                        "com.sun.crypto.provider.AESCipher$General", "com.sun.crypto.provider.ARCFOURCipher",
-                        "com.sun.crypto.provider.ChaCha20Cipher$ChaCha20Poly1305", "com.sun.crypto.provider.DESCipher",
-                        "com.sun.crypto.provider.DESedeCipher", "com.sun.crypto.provider.DHParameters",
-                        "com.sun.crypto.provider.HmacCore$HmacSHA256", "com.sun.management.GarbageCollectorMXBean",
-                        "com.sun.management.GcInfo", "com.sun.management.OperatingSystemMXBean",
-                        "com.sun.management.internal.OperatingSystemImpl", "com.sun.management.ThreadMXBean",
+                        "com.sun.management.GarbageCollectorMXBean",
+                        "com.sun.management.OperatingSystemMXBean",
+                        "com.sun.management.ThreadMXBean",
                         "com.sun.management.UnixOperatingSystemMXBean", "java.lang.management.CompilationMXBean",
                         "java.lang.management.GarbageCollectorMXBean", "java.lang.management.MemoryMXBean",
                         "java.lang.management.MemoryManagerMXBean", "java.lang.management.MemoryPoolMXBean",
                         "java.lang.management.RuntimeMXBean", "java.lang.management.BufferPoolMXBean",
-                        "java.lang.management.ClassLoadingMXBean", "java.lang.management.PlatformLoggingMXBean",
-                        "javax.management.remote.rmi.RMIConnection", "com.sun.management.VMOption",
-                        "javax.management.remote.rmi.RMIConnectionImpl_Stub", "javax.management.remote.rmi.RMIServer",
-                        "javax.management.remote.rmi.RMIServerImpl_Stub", "sun.rmi.registry.RegistryImpl_Stub",
-                        "java.rmi.MarshalledObject", "java.rmi.Remote", "java.rmi.dgc.Lease", "java.rmi.dgc.VMID",
-                        "java.rmi.registry.Registry", "java.rmi.server.ObjID", "java.rmi.server.RemoteObject",
-                        "java.rmi.server.RemoteStub", "java.rmi.server.UID", "javax.management.openmbean.OpenType",
-                        "java.lang.management.LockInfo", "java.lang.management.ManagementPermission",
-                        "java.lang.management.MemoryUsage", "java.lang.management.MonitorInfo",
-                        "java.lang.management.ThreadInfo", "java.security.SecureRandomParameters",
-                        "javax.management.MBeanServerBuilder", "javax.management.NotificationBroadcaster",
-                        "javax.management.NotificationEmitter", "javax.management.NotificationFilterSupport",
-                        "javax.management.ObjectName", "jdk.management.jfr.ConfigurationInfo",
-                        "jdk.management.jfr.EventTypeInfo", "jdk.management.jfr.FlightRecorderMXBean",
-                        "jdk.management.jfr.RecordingInfo", "jdk.management.jfr.SettingDescriptorInfo",
-                        "sun.rmi.registry.RegistryImpl_Skel", "sun.rmi.transport.DGCImpl_Skel",
-                        "sun.rmi.server.UnicastRef2", "sun.rmi.transport.DGCImpl", "sun.rmi.transport.DGCImpl_Skel",
-                        "sun.rmi.transport.DGCImpl_Stub"
+                        "java.lang.management.ClassLoadingMXBean", "java.lang.management.PlatformLoggingMXBean"
         };
 
         String[] methods = new String[]{
-                        "com.sun.management.GcInfo", "com.sun.management.OperatingSystemMXBean",
+                        "com.sun.management.OperatingSystemMXBean",
                         "com.sun.management.ThreadMXBean", "com.sun.management.UnixOperatingSystemMXBean",
-                        "com.sun.management.VMOption", "java.lang.management.BufferPoolMXBean",
+                        "java.lang.management.BufferPoolMXBean",
                         "java.lang.management.ClassLoadingMXBean", "java.lang.management.CompilationMXBean",
                         "java.lang.management.GarbageCollectorMXBean", "java.lang.management.MemoryMXBean",
                         "java.lang.management.MemoryManagerMXBean", "java.lang.management.MemoryPoolMXBean",
-                        "java.lang.management.MemoryUsage", "java.lang.management.RuntimeMXBean",
-                        "java.lang.management.PlatformLoggingMXBean", "java.rmi.registry.Registry",
-                        "javax.management.remote.rmi.RMIConnection", "javax.management.remote.rmi.RMIConnectionImpl_Stub",
-                        "javax.management.remote.rmi.RMIServer", "javax.management.remote.rmi.RMIServerImpl_Stub",
-                        "jdk.management.jfr.FlightRecorderMXBean", "java.lang.management.MonitorInfo",
-                        "java.lang.management.ThreadInfo", "jdk.management.jfr.ConfigurationInfo",
-                        "jdk.management.jfr.EventTypeInfo", "jdk.management.jfr.RecordingInfo",
-                        "jdk.management.jfr.SettingDescriptorInfo", "sun.rmi.registry.RegistryImpl_Stub",
-                        "sun.rmi.server.UnicastRef2", "sun.rmi.transport.DGCImpl", "sun.rmi.transport.DGCImpl_Skel",
-                        "sun.rmi.transport.DGCImpl_Stub"
-        };
-
-        String[] fields = new String[]{"com.sun.management.GcInfo"};
-
-        String[] constructors = new String[]{
-                        "com.sun.management.internal.GarbageCollectorExtImpl",
-                        "com.sun.management.internal.OperatingSystemImpl", "java.lang.management.ManagementPermission",
-                        "javax.management.MBeanServerBuilder", "javax.management.remote.rmi.RMIConnectionImpl_Stub",
-                        "javax.management.remote.rmi.RMIServerImpl_Stub", "sun.rmi.transport.DGCImpl_Stub",
-                        "sun.rmi.registry.RegistryImpl_Skel", "sun.rmi.registry.RegistryImpl_Stub",
-                        "sun.rmi.server.UnicastRef2", "sun.rmi.transport.DGCImpl_Skel"
+                        "java.lang.management.RuntimeMXBean",
+                        "java.lang.management.PlatformLoggingMXBean"
         };
 
         for (String clazz : classes) {
@@ -325,12 +139,6 @@ public class JmxServerFeature implements InternalFeature {
         }
         for (String clazz : methods) {
             RuntimeReflection.register(access.findClassByName(clazz).getMethods());
-        }
-        for (String clazz : constructors) {
-            RuntimeReflection.register(access.findClassByName(clazz).getConstructors());
-        }
-        for (String clazz : fields) {
-            RuntimeReflection.register(access.findClassByName(clazz).getFields());
         }
     }
 }
